Split modules into files
