+++ src/constants.js
@@ -1,0 +1,14 @@
+export const STEP_CONTACT_INFORMATION = 'contact_information';
+export const STEP_SHIPPING_METHOD = 'shipping_method';
+export const STEP_PAYMENT_METHOD = 'payment_method';
+export const STEP_REVIEW = 'review';
+export const STEP_THANK_YOU = 'thank_you';
+export const STEP_ORDER_STATUS = 'order_status';
+export const STEPS_ALL = [
+  STEP_CONTACT_INFORMATION,
+  STEP_SHIPPING_METHOD,
+  STEP_PAYMENT_METHOD,
+  STEP_REVIEW,
+  STEP_THANK_YOU,
+  STEP_ORDER_STATUS
+];
